requirements.txt
pillow
requests
jupyter
numpy
pandas
matplotlib
seaborn
graphviz
Cython
pycocotools
scikit-learn
torch
torchvision
Pillow
opencv-python
transformers
accelerate>=0.26.0
datasets
albumentations
flask
dotenv
optuna